+++ src/PrintC.ml
@@ -82,10 +82,10 @@ and prec_of_op2 op =
   let open Constant in
   match op with
   | AddW | SubW | MultW | DivW -> failwith "[prec_of_op]: should've been desugared"
-  | Add -> 4, 4, 4
+  | Add -> 4, 4, 4 (* addition is associative in C *)
   | Sub -> 4, 4, 3
   | Div -> 3, 3, 2
-  | Mult -> 3, 3, 3
+  | Mult -> 3, 3, 2
   | Mod -> 3, 3, 2
   | BOr -> 10, 10, 10
   | BAnd ->
@@ -93,7 +93,7 @@ and prec_of_op2 op =
   | BXor | Xor -> 9, 9, 9
   | BShiftL | BShiftR ->
       5, 5, 4
-  | Eq | Neq -> 7, 7, 7
+  | Eq | Neq -> 7, 7, 6
   | Lt | Lte | Gt | Gte -> 6, 6, 5
   | And -> 11, 11, 11
   | Or -> 12, 12, 12