+++ src/SimplifyMerge.ml
@@ -70,6 +70,7 @@ let rec merge' (env: env) (u: S.t) (e: expr): S.t * S.t * expr =
   | EString _
   | EAny
   | EOp _
+  | EPolyComp _
   | EPushFrame
   | EPopFrame
   | EEnum _