+++ test/helpers-Server.c
@@ -4,10 +4,6 @@
 
 #include "Server.h"
 
-char char_of_uint8(uint8_t x) {
-  return (char) x;
-}
-
 uint32_t bufstrcpy(char *dst, const char *src) {
   /* The F* precondition guarantees that src is zero-terminated */
   return sprintf(dst, "%s", src);